refactor: move history event tile
